Move initialize_llm into utils.py
